src/requirements.txt
Django==1.11.29
django-bootstrap-form==3.2.1
django-cors-headers==2.1.0
django-markdownx==2.0.28
django-recaptcha==1.3.0
djangorestframework==3.9.1
gunicorn==19.7.1
Markdown==3.0.1
mysqlclient==1.3.10
paymentwall-python==1.0.7
Pillow==5.3.0
pytz==2017.3
PyYAML>=4.2b1
pyjwt==1.7.1
